Exercise 1.37 - continued fractions.
I actually enjoyed this exercise, from the "translation" of the
fractions to Unicode symbols, to writing and testing "cont-frac"
itself.

I especially liked how clearly this exercise shows the difference
between recursive and iterative processes. You can see that the
recursive process is straightforward translation of the fraction in
Scheme, and it's direction is top-down, while the iterative process
goes bottom-up. Nice!