Exercise 1.39 - approximation of tan(x) with continued fractions.
Nothing exiting here.

It's worth nothing that for small values of x (x ∈ (-2π, 2π), for
example) this approximation is very accurate, even for small values of
k, for k = 16 we have accuracy as high as around 10 decimal places!
But when tested with larger values of x (I'm not sure if this makes
sense anyway) it needs large values of k, to achieve acceptable
accuracy. I'm curious why.