Exercise 1.45 - n-th roots with "fixed-point", "average-damp" and "repeated".
I didn't post the results from my experiments. I also don't have a
proof for my idea, although it "seems to work".

This exercise was very interesting. It was also nice to see the power
of the abstractions, comparing the "n-th-root" procedure with the
previous, more specific versions for square and cube root.