Exercise 2.3 - rectangle representation and abstraction barriers.
Interesting, I'm not sure if I got this right, but I'll watch the
video lectures for more insight. I tend to like the second
implementation more than the first, since it's looking more simple - I
think that the procedures to reproduce the corners of the rectangle
out of it's two side will be more complex, than the ones, that create
the sides out of the opposite corners.