Exercise 2.12 - "make-center-percent"
It's cool how you can define different "representations" of one object
just by defining new interface to it. Here we have
"make-center-percent", which uses "make-center-width", which uses
"make-interval". So, although the actual representation is the same (a
pair of numbers), it looks like we have 3 different representations. I
suppose this is "abstraction with data".

It's also worth noting the rounding errors, ("Close enough" :-) ),
maybe this will have something to do with the next exercises.