Exercise 2.29 - binary mobiles
Very good exercise! It was a bit tedious to write the tests, but it
was fun as a whole.

When I was writing "balanced?" I abstracted the procedure
"structure-weight", then I realized, that it could also be used in
"total-weight" itself, using mutual recursion. This reduced much of
the code repetitions. Then I made the "simple-weight?" procedure,
which further simplified the logic behind "balanced?".

Furthermore, these two helper procedures put another abstraction
barrier between the procedures "total-weight" and "balanced?", and the
internal representation of mobiles and branches. Should we change the
way we represent mobiles and branches (by coding them with natural
numbers, for example, as in ex. 2.5), we should rewrite the
constructors and selectors, of course, but beyond that we must rewrite
only "structure-weight" and "simple-weight?", leaving the code of
"total-weight" and "balanced?" intact. I find this very elegant.